HTML, CSS, JS - Switched to JS
Switched to js instead of ts

HTML
Removed font-awesome - Didn't use it

Button's "data-tab" -> "data-for-tab"
Content -> "data-tab"

Put the elements in the cards in a div class.

CSS
Removed unnecessary css styles

Changed heading text margins

Added half-press confirmation to tabs.
Changed colors when tab is selected.
Made the selection animation not play when a tab is already selected.
Added a selected image when tab is selected.

Added thumbnail height to the cards for eventual thumbnails.

Added @media dimensions to take care of the heading, divider, and the text in the tab-button to stay on 1 line.

JS
Added tab switch function